+++ database/migrations/20200626090432_add_parent_to_infos_table.php
@@ -1,0 +1,25 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use App\Info;
+
+class AddParentToInfosTable extends Migration
+{
+    public function up()
+    {
+        Info::truncate();
+
+        $this->schema->table('infos', function (Blueprint $table) {
+            $table->string('parent')->nullable();
+            $table->index('parent');
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('infos', function (Blueprint $table) {
+            $table->dropColumn('parent');
+        });
+    }
+}